chore: Rename randomUtils to apiTestUtils
